Extract Document Graph canvas rendering to separate module
Move canvas painting functions (node rendering, icons, links) from MindMap.tsx to a new mindMapCanvas.ts module. This separates rendering logic from interaction handling and ensures the painted glyph and click hit test remain synchronized. Includes updating test imports.
